+++ kuro/client/clients.py
@@ -1,9 +1,15 @@
 """Main client for requesting API endpoints."""
 
-from kuro.client.components import announcement, auth, gacha
+from kuro.client.components import announcements, auth, gacha
 
 __all__ = ["Client"]
 
 
-class Client(auth.AuthClient, gacha.GachaClient, announcement.AnnouncementClient):
+class Client(
+    auth.AuthClient,
+    gacha.GachaClient,
+    announcements.KurobbsAnnouncementClient,
+    announcements.GameAnnouncementClient,
+    announcements.LauncherAnnouncementClient,
+):
     """Main client for requesting API endpoints."""
