+++ kuro/client/components/announcements/__init__.py
@@ -1,0 +1,5 @@
+"""Announcement Clients."""
+
+from .game_announcement import *
+from .kurobbs_announcement import *
+from .launcher_announcement import *
