+++ kuro/models/__init__.py
@@ -1,6 +1,6 @@
 """Pydantic models used by the library."""
 
-from .announcement import *
+from .announcements import *
 from .assets import *
 from .auth import *
 from .base import *
