+++ kuro/models/announcements/__init__.py
@@ -1,0 +1,5 @@
+"""Announcement models."""
+
+from .game_announcement import *
+from .kurobbs_announcement import *
+from .launcher_announcement import *
